Crate ggml_gpt_oss_20b_GGUF_curl_naive_strip_tags

Naive stripping html tags (not working for CSS etc)

rustc ggml-gpt-oss-20b-GGUF-curl-naive-strip_tags.rs && ./ggml-gpt-oss-20b-GGUF-curl-naive-strip_tags| tee ggml-gpt-oss-20b-GGUF-curl-naive-strip_tags.log

html

<!DOCTYPE html>
<html lang="en">
<head>
<title>curl</title>
<meta name="viewport" content="width=device-width, initial-scale=1.0">
<meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
<link rel="stylesheet" type="text/css" href="/curl.css">
<link rel="shortcut icon" href="/favicon.ico">
<link rel="icon" href="/logo/curl-symbol.svg" type="image/svg+xml">
<link rel="alternate" type="application/rss+xml" title="curl Releases" href="https://github.com/curl/curl/releases.atom">
<style type="text/css">
.sponsors {
  padding: 5px 5px 5px 5px;
  border: 2px black solid;
  border-radius: 10px;
  margin-bottom: 20px;
  text-align: center;
  background: #e0e0e0;
}
.logo {
  max-width: 80%;
  margin: 20px 0px 0px 0px;
}
.sponsors a:hover, .logo:hover {
  background: transparent;
}
.sidebar-box {
  padding: 5px;
  margin-bottom: 20px;
  border: 2px black solid;
  border-radius: 10px;
  background: #e0e0e0;
}
td {
  vertical-align: top;
}
.left {
  font-weight: bold;
  padding: 4px 4px 4px 4px;
  font-size: 110%;
}
.right {
  padding: 4px 4px 4px 4px;
}
.l {
  background: #093754;
  border: 0px;
  padding: 15px;
}
.r {
  background: #0f564d;
  border: 0px;
}
.supports {
  color: #f0f0f0;
  font-size: 90%;
  padding: 0 0 0 0;
}
@media (prefers-color-scheme: dark) {
  .sidebar-box {
    border-color: #999;
    background: #404040;
  }
  .sponsors {
    border-color: #999;
    background: #404040;
  }
}
</style>
</head>
<body>
<div class="main">
<div class="menu">
<div class="curlsymbol">
  <a href="/">
    <img src="/logo/curl-white-symbol.svg">
  </a>
</div>
<div class="dropdown">
  <a class="dropbtn" href="/download.html">Download</a>
  <div class="dropdown-content">
    <a href="https://github.com/curl/curl">Browse source</a>
    <a href="/ch/">Changelog</a>
    <a href="/rc/">Release candidates</a>
    <a href="/docs/releases.html">Release log</a>
    <a href="/tiny/">tiny-curl</a>
  </div>
</div>
<div class="dropdown">
  <a class="dropbtn" href="/docs/">Documentation</a>
  <div class="dropdown-content">
    <a href="/docs/projdocs.html">Project</a>
      <small>
      <a href="/docs/faq.html">FAQ</a>
      <a href="/docs/help-us.html">&nbsp; Help us</a>
      <a href="/docs/knownbugs.html">&nbsp; Known bugs</a>
      <a href="/docs/knownrisks.html">&nbsp; Known risks</a>
      <a href="/docs/todo.html">&nbsp; TODO</a>
      </small>
    <a href="/docs/protdocs.html">Protocols</a>
      <small>
      <a href="/docs/caextract.html">&nbsp; CA bundle</a>
      <a href="/docs/http-cookies.html">&nbsp; HTTP Cookies</a>
      <a href="/docs/sslcerts.html">&nbsp; SSL Certs</a>
      </small>
    <a href="/docs/reldocs.html">Releases</a>
      <small>
      <a href="/docs/security.html">&nbsp; Security</a>
      <a href="/docs/versions.html">&nbsp; Version numbers</a>
      <a href="/docs/vulnerabilities.html">&nbsp; Vulnerabilities</a>
      </small>
    <a href="/docs/tooldocs.html">curl tool</a>
      <small>
      <a href="/docs/manpage.html">&nbsp; man page</a>
      <a href="/docs/tutorial.html">&nbsp; Tutorial</a>
      <a href="/docs/httpscripting.html">&nbsp; HTTP scripting</a>
      </small>
    <a href="/trurl/">trurl</a>
    <a href="/wcurl/">wcurl</a>
    <a href="/docs/videos/">Videos</a>
    <a href="/docs/whodocs.html">Who and Why</a>
  </div>
</div>
<div class="dropdown">
  <a class="dropbtn" href="/libcurl/">libcurl</a>
  <div class="dropdown-content">
    <a href="/libcurl/c/">API</a>
    <a href="/libcurl/c/example.html">Examples</a>
    <a href="/libcurl/features.html">Features</a>
    <a href="/mail/list.cgi?list=curl-library">Mailing list</a>
    <a href="/libcurl/c/symbols-in-versions.html">Symbols</a>
    <a href="/libcurl/using/">Using libcurl</a>
    <a href="/libcurl/c/libcurl-tutorial.html">Tutorial</a>
  </div>
</div>
<div class="dropdown">
  <a class="dropbtn" href="/gethelp.html">Get Help</a>
  <div class="dropdown-content">
    <a href="https://lists.haxx.se/listinfo/curl-library">curl-library</a>
    <a href="https://lists.haxx.se/listinfo/curl-users">curl-users</a>
    <a href="/docs/irc.html">IRC / chat</a>
    <a href="/mail/">Mailing lists</a>
    <a href="/book.html">Everything curl [book]</a>
    <a href="/docs/videos/">Video presentations</a>
    <a href="https://github.com/curl/curl/issues">Report a bug</a>
    <a href="/dev/vuln-disclosure.html">Report security issue</a>
    <a href="/support.html">Paid support</a>
  </div>
</div>
<div class="dropdown">
  <a class="dropbtn" href="/dev/">Development</a>
  <div class="dropdown-content">
    <a href="/dev/builds.html">Autobuilds</a>
    <a href="/dev/code-review.html">Code review</a>
    <a href="/dev/code-style.html">Code style</a>
    <a href="/dev/contribute.html">Contribute</a>
    <a href="/dashboard.html">Dashboard</a>
    <a href="/dev/deprecate.html">Deprecate</a>
    <a href="/dev/internals.html">Internals</a>
    <a href="/dev/release-notes.html">Release Notes</a>
    <a href="/dev/release-procedure.html">Release Procedure</a>
    <a href="/dev/roadmap.html">Roadmap</a>
    <a href="/dev/runtests.html">Run Tests</a>
    <a href="/rfc/">Specifications</a>
    <a href="/dev/testcurl.html">Test curl</a>
    <a href="/dev/tests-overview.html">Tests Overview</a>
    <a href="/dev/vuln-disclosure.html">Vulnerability Disclosure Policy</a>
  </div>
</div>
</div>
<div class="contents">
<center><img src="https://curl.se/logo/curl-logo.svg" alt="curl logo" style="width:90%; padding-top: 8px;">
</center>
<p>
<hr>
<p style="font-size: 160%; text-align: center;">
  <b>command line tool and
library</b>
  <br>for transferring data with URLs
  <br>(since 1998)</p>
<div style="float: right; max-width: 10em; margin-left: 40px;">
  <div class="sponsors">
  <b>Top&nbsp;Sponsors</b>
  <a href="sponsors.html">
  <img class="logo" src="pix/haxx.svg" alt="Haxx">
  <img class="logo" src="pix/wolfssl.svg" alt="wolfSSL">
  <img class="logo" src="pix/fastly.svg" alt="Fastly">
  <img class="logo" src="pix/top/github.svg" alt="GitHub">
  <img class="logo" src="pix/teamviewer.svg" alt="Teamviewer">
  </a>
 </div>
 <div class="sidebar-box">
   <img style="filter: grayscale(100%); float: right;" src="pix/user.svg" width="30%">
   <a href="/support.html">
   Commercial support available</a>
   <div style="clear: both;"></div>
 </div>
 <div class="sidebar-box">
  <a href="https://everything.curl.dev/">Everything curl</a> is a free ebook
  about curl and libcurl
  <p align="center">
  <a href="https://everything.curl.dev/"><img src="pix/everything-curl.jpg"
  alt="Everything curl" width="80%"></a>
 </div>
 <div class="sidebar-box">
   <img style="filter: grayscale(100%); float: right;" src="pix/cash.svg" width="40%">
   <a href="/donation.html">
   Sponsor curl financially</a>
   <div style="clear: both;"></div>
 </div>
 <div class="sidebar-box">
   Video: <a href="https://youtu.be/V5vZWHP-RqU">mastering the curl command line tool</a>
   <p align="center">
   <a href="https://youtu.be/V5vZWHP-RqU"><img src="pix/mastering.jpg" width="80%"></a>
 </div>
</div>
<h2> What is curl used for? </h2>
<p>
curl is used in command lines or scripts to transfer data. curl is
also <b>libcurl</b>, used in cars, television sets, routers, printers, audio
equipment, mobile phones, tablets, medical devices, settop boxes, computer
games, media players and is the Internet transfer engine for countless
software applications in over <i>twenty billion installations</i>.
<p>
curl is used daily by virtually every Internet-using human on the globe.
<h2> curl supports </h2>
<table class="supports" cellpadding="0" cellspacing="0">
<tr class="l"><td class="left">Protocols</td>
  <td class="right"> DICT, FILE, FTP, FTPS, GOPHER, GOPHERS, HTTP, HTTPS,
IMAP, IMAPS, LDAP, LDAPS, MQTT, MQTTS, POP3, POP3S, RTMP, RTMPS, RTSP, SCP,
SFTP, SMB, SMBS, SMTP, SMTPS, TELNET, TFTP, WS, WSS </td></tr>
<tr class="r"><td class="left">Proxies</td>
<td class="right"> SOCKS4, SOCKS5, HTTP, HTTPS (HTTP/1 and HTTP/2), tunneling,
via unix domain sockets, haproxy, SOCKS+HTTP proxy chain</td></tr>
<tr class="l"><td class="left">HTTP</td>
<td class="right">GET, POST, PUT, HEAD, multipart formpost, HTTP/0.9,
HTTP/1.0, HTTP/1.1, HTTP/2 (h2c, h2, prior knowledge), HTTP/3 (dual connect
h1/h2 + h3 or h3-only), HSTS, Alt-Svc, cookies, PSL, etags, transfer
compression, ranges, custom headers, custom method, follow redirects</td></tr>
<tr class="r"><td class="left">FTP</td>
<td class="right">IPv6 (EPRT, EPSV), STLS, upload/download, append, range,
passive/active, kerberos, directory listing, custom commands</td></tr>
<tr class="l"><td class="left">SCP + SFTP</td>
<td class="right">known hosts, md5/sha256 fingerprint, compression,
upload/download, directory listing</td></tr>
<tr class="r"><td class="left">TLS</td>
<td class="right"> 1.0 - 1.3, mutual authentication, STARTTLS, OCSP stapling,
ECH, False Start, key pinning, PQC ready, session resumption, early data,
export/import sessions</td></tr>
<tr class="l"><td class="left">QUIC</td>
  <td class="right">
0RTT handshakes
</td></tr>
<tr class="r"><td class="left">Auth</td>
<td class="right">Basic, Plain, Digest, CRAM-MD5, SCRAM-SHA, NTLM, Negotiate,
Kerberos, Bearer tokens, AWS Sigv4, SASL, .netrc</td></tr>
<tr class="l"><td class="left">HTTP&nbsp;compression</td>
<td class="right">gzip, brotli and zstd</td></tr>
<tr class="r"><td class="left">Name resolving</td>
<td class="right">DNS-over-HTTPS, custom address for host, name+port redirect,
custom DNS servers, DNS caching, HTTPS RR</td></tr>
<tr class="l"><td class="left">Connection</td>
<td class="right">connection reuse, Interface binding, Happy Eyeballs,
IPv4/IPv6-only, unix domain sockets, TCP keepalive, TCP Fast Open, TCP
Nodelay, MPTCP, VLAN priority, IP Type Of Service</td></tr>
<tr class="r"><td class="left">Transfers</td>
<td class="right">transfer rate limiting, request rate limiting, stall
detection, retries, timeouts</td></tr>
<tr class="l"><td class="left">URLs</td>
<td class="right">Unlimited amount, parallel and serial transfers,
globbing</td></tr>
<tr class="r"><td class="left">Output</td>
<td class="right">IDN hostnames, custom info from transfer, metadata as JSON,
per content-disposition, libcurl source code, bold headers</td></tr>
</table>
<h2> Who makes curl? </h2>
<p>
curl is free and <a href="docs/copyright.html" title="curl and libcurl are
licensed under a MIT/X derivate license">open source</a> software and exists
thanks to <a href="docs/thanks.html" title="over 3,000 named
contributors">thousands of contributors</a> and our
awesome <a href="/sponsors.html">sponsors</a>. The curl
project <a title="Adheres to the Core Intrastructure Initiative's listed best
practices at Gold level"
href="https://bestpractices.coreinfrastructure.org/projects/63">follows well
established open source best practices</a>. You too
can <a href="docs/help-us.html">help us</a> improve!
<h2> What is the latest curl? </h2>
<p>
The most recent stable version is <b>8.18.0</b>, released on 2026-01-07.
Currently, 102 of the listed <a href="download.html" title="Binary download packages for your platform!">downloads</a> are of the latest version.
<h2> Where is the code? </h2>
<p>
 Check out the latest <a href="https://github.com/curl/curl" title="git clone
 the curl code">source code from GitHub</a>.
</div>
</div>
</body>
</html>

text

curl

.sponsors {
  padding: 5px 5px 5px 5px;
  border: 2px black solid;
  border-radius: 10px;
  margin-bottom: 20px;
  text-align: center;
  background: #e0e0e0;
}
.logo {
  max-width: 80%;
  margin: 20px 0px 0px 0px;
}
.sponsors a:hover, .logo:hover {
  background: transparent;
}
.sidebar-box {
  padding: 5px;
  margin-bottom: 20px;
  border: 2px black solid;
  border-radius: 10px;
  background: #e0e0e0;
}
td {
  vertical-align: top;
}
.left {
  font-weight: bold;
  padding: 4px 4px 4px 4px;
  font-size: 110%;
}
.right {
  padding: 4px 4px 4px 4px;
}
.l {
  background: #093754;
  border: 0px;
  padding: 15px;
}
.r {
  background: #0f564d;
  border: 0px;
}
.supports {
  color: #f0f0f0;
  font-size: 90%;
  padding: 0 0 0 0;
}
@media (prefers-color-scheme: dark) {
  .sidebar-box {
    border-color: #999;
    background: #404040;
  }
  .sponsors {
    border-color: #999;
    background: #404040;
  }
}

  
    
  

  Download
  
    Browse source
    Changelog
    Release candidates
    Release log
    tiny-curl
  

  Documentation
  
    Project
      
      FAQ
      &nbsp; Help us
      &nbsp; Known bugs
      &nbsp; Known risks
      &nbsp; TODO
      
    Protocols
      
      &nbsp; CA bundle
      &nbsp; HTTP Cookies
      &nbsp; SSL Certs
      
    Releases
      
      &nbsp; Security
      &nbsp; Version numbers
      &nbsp; Vulnerabilities
      
    curl tool
      
      &nbsp; man page
      &nbsp; Tutorial
      &nbsp; HTTP scripting
      
    trurl
    wcurl
    Videos
    Who and Why
  

  libcurl
  
    API
    Examples
    Features
    Mailing list
    Symbols
    Using libcurl
    Tutorial
  

  Get Help
  
    curl-library
    curl-users
    IRC / chat
    Mailing lists
    Everything curl [book]
    Video presentations
    Report a bug
    Report security issue
    Paid support
  

  Development
  
    Autobuilds
    Code review
    Code style
    Contribute
    Dashboard
    Deprecate
    Internals
    Release Notes
    Release Procedure
    Roadmap
    Run Tests
    Specifications
    Test curl
    Tests Overview
    Vulnerability Disclosure Policy
  

  command line tool and
library
  for transferring data with URLs
  (since 1998)

  
  Top&nbsp;Sponsors
  
  
  
  
  
  
  
 
 
   
   
   Commercial support available
   
 
 
  Everything curl is a free ebook
  about curl and libcurl
  
  
 
 
   
   
   Sponsor curl financially
   
 
 
   Video: mastering the curl command line tool
   
   
 

 What is curl used for? 

curl is used in command lines or scripts to transfer data. curl is
also libcurl, used in cars, television sets, routers, printers, audio
equipment, mobile phones, tablets, medical devices, settop boxes, computer
games, media players and is the Internet transfer engine for countless
software applications in over twenty billion installations.

curl is used daily by virtually every Internet-using human on the globe.
 curl supports 

Protocols
   DICT, FILE, FTP, FTPS, GOPHER, GOPHERS, HTTP, HTTPS,
IMAP, IMAPS, LDAP, LDAPS, MQTT, MQTTS, POP3, POP3S, RTMP, RTMPS, RTSP, SCP,
SFTP, SMB, SMBS, SMTP, SMTPS, TELNET, TFTP, WS, WSS 
Proxies
 SOCKS4, SOCKS5, HTTP, HTTPS (HTTP/1 and HTTP/2), tunneling,
via unix domain sockets, haproxy, SOCKS+HTTP proxy chain
HTTP
GET, POST, PUT, HEAD, multipart formpost, HTTP/0.9,
HTTP/1.0, HTTP/1.1, HTTP/2 (h2c, h2, prior knowledge), HTTP/3 (dual connect
h1/h2 + h3 or h3-only), HSTS, Alt-Svc, cookies, PSL, etags, transfer
compression, ranges, custom headers, custom method, follow redirects
FTP
IPv6 (EPRT, EPSV), STLS, upload/download, append, range,
passive/active, kerberos, directory listing, custom commands
SCP + SFTP
known hosts, md5/sha256 fingerprint, compression,
upload/download, directory listing
TLS
 1.0 - 1.3, mutual authentication, STARTTLS, OCSP stapling,
ECH, False Start, key pinning, PQC ready, session resumption, early data,
export/import sessions
QUIC
  
0RTT handshakes

Auth
Basic, Plain, Digest, CRAM-MD5, SCRAM-SHA, NTLM, Negotiate,
Kerberos, Bearer tokens, AWS Sigv4, SASL, .netrc
HTTP&nbsp;compression
gzip, brotli and zstd
Name resolving
DNS-over-HTTPS, custom address for host, name+port redirect,
custom DNS servers, DNS caching, HTTPS RR
Connection
connection reuse, Interface binding, Happy Eyeballs,
IPv4/IPv6-only, unix domain sockets, TCP keepalive, TCP Fast Open, TCP
Nodelay, MPTCP, VLAN priority, IP Type Of Service
Transfers
transfer rate limiting, request rate limiting, stall
detection, retries, timeouts
URLs
Unlimited amount, parallel and serial transfers,
globbing
Output
IDN hostnames, custom info from transfer, metadata as JSON,
per content-disposition, libcurl source code, bold headers

 Who makes curl? 

curl is free and open source software and exists
thanks to thousands of contributors and our
awesome sponsors. The curl
project follows well
established open source best practices. You too
can help us improve!
 What is the latest curl? 

The most recent stable version is 8.18.0, released on 2026-01-07.
Currently, 102 of the listed downloads are of the latest version.
 Where is the code? 

 Check out the latest source code from GitHub.